refactor tests to center around each collection
