+++ package.json
@@ -14,8 +14,8 @@
     "debug": "cross-env DEBUG=Eleventy* pnpm exec eleventy",
     "debugstart": "cross-env DEBUG=Eleventy* pnpm exec eleventy --serve --quiet",
     "benchmark": "cross-env DEBUG=Eleventy:Benchmark* pnpm exec eleventy",
-    "test": "pnpm exec jest --experimental-vm-modules --detectOpenHandles",
-    "test:watch": "pnpm exec jest --experimental-vm-modules --watch --detectOpenHandles",
+    "test": "NODE_OPTIONS=--experimental-vm-modules pnpm exec jest --detectOpenHandles",
+    "test:watch": "NODE_OPTIONS=--experimental-vm-modules pnpm exec jest --watch --detectOpenHandles",
     "test:directus": "pnpm test -- -t 'Directus'",
     "setup": "./setup-dev.sh",
     "directus:setup": "cd tools/directus && ./setup-dev-environment.sh",
@@ -40,5 +40,19 @@
     "@babel/preset-env": "^7.26.9",
     "jest": "^29.7.0",
     "jest-environment-node": "^29.7.0"
+  },
+  "jest": {
+    "testEnvironment": "node",
+    "transform": {},
+    "moduleNameMapper": {
+      "^(\\.{1,2}/.*)\\.js$": "$1"
+    },
+    "testMatch": ["**/tests/**/*.test.js"],
+    "verbose": true,
+    "globalSetup": "./tests/setup.js",
+    "globalTeardown": "./tests/teardown.js",
+    "transformIgnorePatterns": [
+      "node_modules/(?!(@directus)/)"
+    ]
   }
 }
