+++ tests/collections/projects.test.js
@@ -14,6 +14,8 @@ import { projects } from '../../src/directus/collections/index.js';
 
 describe('projects Collection', () => {
   let client;
+  let token;
+  const { url } = getDirectusConfig();
   
   beforeAll(async () => {
     // Create Directus client
@@ -22,6 +24,9 @@ describe('projects Collection', () => {
     // Login to Directus
     const loggedIn = await loginToDirectus(client);
     expect(loggedIn).toBe(true);
+    
+    // Get token for API calls
+    token = client.getToken();
   });
   
   test('Collection exists and is accessible', async () => {
@@ -33,10 +38,9 @@ describe('projects Collection', () => {
       expect(Array.isArray(items)).toBe(true);
       
       // Get the total count using the API directly
-      const { url } = getDirectusConfig();
       const response = await fetch(`${url}/items/projects?limit=1`, {
         headers: {
-          'Authorization': `Bearer ${client.getToken()}`
+          'Authorization': `Bearer ${token}`
         }
       });
       
@@ -56,16 +60,71 @@ describe('projects Collection', () => {
     }
   });
   
-  test('Collection has the expected structure', async () => {
-    // Get the expected fields from the collection definition
-    const expectedFields = projects.fields.map(field => field.field);
+  describe('Field structure', () => {
+    let fields;
     
-    // Log the expected fields
-    console.log('Collection projects exists and is accessible');
-    console.log('Expected fields:', expectedFields.join(', '));
+    beforeAll(async () => {
+      // Get the expected fields from the collection definition
+      fields = projects.fields.map(field => ({
+        field: field.field,
+        type: field.type
+      }));
+      
+      // Log the expected fields
+      console.log('Expected fields:', fields.map(f => f.field).join(', '));
+    });
     
-    // We could add more detailed tests here to verify the field structure
-    // by fetching the collection schema from Directus, but that would require
-    // admin permissions
+    test('has title field (string)', () => {
+      const field = fields.find(f => f.field === 'title');
+      expect(field).toBeDefined();
+      expect(field.type).toBe('string');
+    });
+    
+    test('has status field (enum)', () => {
+      const field = fields.find(f => f.field === 'status');
+      expect(field).toBeDefined();
+      expect(['string', 'enum'].includes(field.type)).toBe(true);
+    });
+    
+    test('has description field (text)', () => {
+      const field = fields.find(f => f.field === 'description');
+      expect(field).toBeDefined();
+      expect(['text', 'string'].includes(field.type)).toBe(true);
+    });
+    
+    test('has technologies field (json or multiple select)', () => {
+      const field = fields.find(f => f.field === 'technologies');
+      expect(field).toBeDefined();
+      expect(['json', 'csv', 'array'].includes(field.type)).toBe(true);
+    });
+    
+    test('has short_description field (text)', () => {
+      const field = fields.find(f => f.field === 'short_description');
+      expect(field).toBeDefined();
+      expect(['text', 'string'].includes(field.type)).toBe(true);
+    });
+    
+    test('has github_url field (string)', () => {
+      const field = fields.find(f => f.field === 'github_url');
+      expect(field).toBeDefined();
+      expect(field.type).toBe('string');
+    });
+    
+    test('has featured_image field (file optional)', () => {
+      const field = fields.find(f => f.field === 'featured_image');
+      expect(field).toBeDefined();
+      expect(['uuid', 'file', 'integer'].includes(field.type)).toBe(true);
+    });
+    
+    test('has all expected fields', () => {
+      // Get the expected fields from the collection definition
+      const expectedFields = projects.fields.map(field => field.field);
+      
+      // Check that all expected fields are present
+      for (const fieldName of expectedFields) {
+        const field = fields.find(f => f.field === fieldName);
+        expect(field).toBeDefined();
+      }
+    });
   });
 });
